Visibility-Agent/measure/requirements.txt
psutil
distro
py-cpuinfo
kafka-python